requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.2
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.25
cffi==1.9.1
configparser==3.5.0
cryptography==1.7.2
decorator==4.0.11
diff-match-patch==20121119
Django==1.10.5
django-autofixture==0.12.1
django-braces==1.11.0
django-celery==3.2.1
django-crispy-forms==1.6.1
django-extensions==1.7.7
django-filter==1.0.1
django-import-export==0.5.1
djangorestframework==3.5.4
enum34==1.1.6
flake8==3.3.0
funcsigs==1.0.2
future==0.16.0
httplib2==0.10.3
httpretty==0.8.14
idna==2.4
inflection==0.3.1
ipaddress==1.0.18
ipython==5.3.0
ipython-genutils==0.1.0
kombu==3.0.37
Markdown==2.6.8
mccabe==0.6.1
mock==2.0.0
more-itertools==2.5.0
ndg-httpsclient==0.4.2
numpy==1.12.0
oauth2==1.9.0.post1
oauthlib==2.0.1
packaging==16.8
pandas==0.19.2
pathlib2==2.2.1
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
protobuf==3.2.0
ptyprocess==0.5.1
py==1.4.32
pyasn1==0.2.3
pycodestyle==2.3.1
pycparser==2.17
pydotplus==2.0.2
pyflakes==1.5.0
Pygments==2.2.0
pyOpenSSL==16.2.0
pyparsing==2.1.10
pytest==3.0.6
pytest-django==3.1.2
python-dateutil==2.6.0
python-twitter==3.2.1
-e git+https://github.com/jabbalaci/pytumblr.git@07e7ccd76ef045458335299fd17cde56f3cab576#egg=PyTumblr
pytz==2016.10
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
scandir==1.5
scikit-learn==0.18.1
scipy==0.18.1
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
tablib==0.11.4
tensorflow==1.0.0
traitlets==4.3.2
vine==1.1.3
wcwidth==0.1.7
